refactor: move work manager code to a service class
